backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Validation & Settings
pydantic==2.10.6
pydantic-settings==2.7.1
annotated-types==0.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
bcrypt==4.2.1

# Utils
python-dotenv==1.0.1
python-multipart==0.0.20
email-validator==2.2.0
python-dateutil==2.9.0

# Security
cryptography==44.0.0
slowapi==0.1.9

# HTTP
requests==2.32.3
httpx==0.28.1
certifi==2024.8.30

# Email & SMS
python-decouple==3.8
aiosmtplib==3.0.2
twilio==9.4.0

# OAuth2 & Google
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Testing
pytest==8.3.4

# Core dependencies
typing-extensions==4.12.2
anyio==4.6.2
idna==3.10
sniffio==1.3.1

